copy the signature to the upload file
